Squashed commit of the following:
commit 102606e429cd803b15b7ac3087be43c093f5f8f7
Author: brijeshlakkad <brijeshlakkad22@gmail.com>
Date:   Sun Mar 29 14:09:23 2020 +0530

    Base controller and its services fixed. Authority and User relationship fixed

commit 91f1ff42e3031c61cd7790441d74be2a52dc3793
Author: brijeshlakkad <brijeshlakkad22@gmail.com>
Date:   Sat Mar 28 01:55:49 2020 +0530

    comment and it's child relation has been fixed

commit b476f529def05e3de06b78d8e0aa80b75d83f3a9
Author: brijeshlakkad <brijeshlakkad22@gmail.com>
Date:   Fri Mar 27 00:14:22 2020 +0530

    Access control structure defined and used for post and comment. Remaining to add for other services.

    AccessControl structure defined and used for post and comment. Remaining to add for other services.

+++ src/main/java/com/questnr/controllers/community/CommunityController.java
@@ -1,10 +1,12 @@
 package com.questnr.controllers.community;
 
+import com.questnr.exceptions.AccessException;
 import com.questnr.model.dto.CommunityDTO;
 import com.questnr.model.dto.UserDTO;
 import com.questnr.model.entities.Community;
 import com.questnr.model.mapper.CommunityMapper;
 import com.questnr.model.mapper.UserMapper;
+import com.questnr.services.access.CommunityAvatarAccessService;
 import com.questnr.services.community.CommunityCommonService;
 import com.questnr.services.community.CommunityService;
 import org.mapstruct.factory.Mappers;
@@ -16,12 +18,17 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import javax.validation.Valid;
-import java.util.ArrayList;
 import java.util.List;
 
 @RestController
 @RequestMapping(value = "/api/v1")
 public class CommunityController {
+
+    final String errorMessage = "You don't have access for the particular operation";
+
+    @Autowired
+    CommunityAvatarAccessService communityAvatarService;
+
     @Autowired
     CommunityService communityService;
 
@@ -35,14 +42,19 @@ public class CommunityController {
     UserMapper userMapper;
 
     CommunityController() {
-
         communityMapper = Mappers.getMapper(CommunityMapper.class);
     }
 
     // Community CRUD Operations
     @RequestMapping(value = "/community", method = RequestMethod.POST, consumes = {MediaType.MULTIPART_FORM_DATA_VALUE}, produces = {MediaType.APPLICATION_JSON_VALUE})
     CommunityDTO createCommunity(@Valid Community requests, @Nullable @RequestParam(value = "file") MultipartFile multipartFile ) {
-        return communityMapper.toDTO(communityService.createCommunity(requests, multipartFile));
+        /*
+         * Community Creation Security Checking
+         * */
+        if(communityAvatarService.hasAccessToCommunityCreation()) {
+            return communityMapper.toDTO(communityService.createCommunity(requests, multipartFile));
+        }
+        throw new AccessException(errorMessage);
     }
 
     @RequestMapping(value = "/community/{communityId}", method = RequestMethod.GET)
@@ -53,7 +65,14 @@ public class CommunityController {
     @RequestMapping(value = "/community/{communityId}", method = RequestMethod.DELETE)
     @ResponseStatus(HttpStatus.OK)
     void deleteCommunity(@PathVariable long communityId) {
-        communityService.deleteCommunity(communityId);
+        /*
+         * Community Deletion Security Checking
+         * */
+        if(communityAvatarService.hasAccessToCommunityDeletion()) {
+            communityService.deleteCommunity(communityId);
+        }else{
+            throw new AccessException(errorMessage);
+        }
     }
 
     // Get users of a single community.
